+++ src/screens/ApplyScreen/ApplyForm/BasicInformationForm/index.tsx
@@ -1,5 +1,3 @@
-import { useFormContext } from 'react-hook-form';
-import { BasicInformationSchemaType } from '../../schemas';
 import {
   FormControl,
   FormField,
@@ -14,121 +12,22 @@ import { Select, SelectTrigger, SelectValue, SelectContent } from '@/components/
 import { Year } from '@/components/ui/Select/Birthday/year';
 import { Month } from '@/components/ui/Select/Birthday/month';
 import { Day } from '@/components/ui/Select/Birthday/day';
-import { getDaysInMonth } from '@/utils/days';
 import { Prefecture } from '@/components/ui/Select/Prefecture';
 import { CityWrapper } from '@/screens/ApplyScreen/ApplyForm/BasicInformationForm/CityWrapper';
-import { getZipcodeOrList } from '@/__generated_REST__/zipcloud/zipcloud';
+import { useBasicInformaionForm } from '@/screens/ApplyScreen/ApplyForm/BasicInformationForm/useBasicInformationForm';
 
 export const BasicInfomationForm = () => {
-  const { control, watch, setValue, trigger } = useFormContext<BasicInformationSchemaType>();
-
-  //
-  // TODO: hooksでまとめたいけど、わからない
-  //
-
-  const currentYear = new Date().getFullYear();
-  const startYear = currentYear - 17;
-  const endYear = 1900;
-
-  const selectedYear = watch('year');
-  const selectedMonth = watch('month');
-  const selectedDay = watch('day');
-
-  const selectedPrefecture = watch('prefectureId');
-  //
-  // TODO: 外部APIの都合上、自動反映を後回し（watchしているのでリストは作成される）
-  //
-  // const selectedCity = watch('cityId');
-
-  //
-  // NOTE: 選択された年をフォームの 'year' フィールドに設定する
-  //       年月日の状態によって、日を空にするため、onChange時に拡張している
-  //
-  const handleYearChange = (value: string) => {
-    // 現在選択されている `日` が、選択されたYear, Monthでの日に存在しない場合
-    if (getDaysInMonth(value, selectedMonth) < Number(selectedDay)) {
-      setValue('day', '');
-    }
-    setValue('year', value);
-
-    //
-    // TODO: birthdayというキーでエラーを管理しているので、
-    //       生年月日のonChangeで毎回手動でバリデーションチェックする
-    //       refineのpathがbirthdayなので、こういう書き方しかできないかもしれない
-    //
-    // @ts-ignore
-    trigger('birthday');
-  };
-
-  //
-  // NOTE: 選択された月をフォームの 'month' フィールドに設定する
-  //       年月日の状態によって、日を空にするため、onChange時に拡張している
-  //
-  const handleMonthChange = (value: string) => {
-    // 現在選択されている `日` が、選択されたYear, Monthでの日に存在しない場合
-    if (getDaysInMonth(selectedYear, value) < Number(selectedDay)) {
-      setValue('day', '');
-    }
-    setValue('month', value);
-
-    //
-    // TODO: birthdayというキーでエラーを管理しているので、
-    //       生年月日のonChangeで毎回手動でバリデーションチェックする
-    //       refineのpathがbirthdayなので、こういう書き方しかできないかもしれない
-    //
-    // @ts-ignore
-    trigger('birthday');
-  };
-
-  const handleDayChange = (value: string) => {
-    setValue('day', value);
-    //
-    // TODO: birthdayというキーでエラーを管理しているので、
-    //       生年月日のonChangeで毎回手動でバリデーションチェックする
-    //       refineのpathがbirthdayなので、こういう書き方しかできないかもしれない
-    //
-    // @ts-ignore
-    trigger('birthday');
-  };
-
-  const handlePostalCodeChange = async (value: string) => {
-    setValue('postalCode', value);
-
-    if (value.length === 7) {
-      //
-      // TODO: refineで定義する必要があるか問題
-      //       onChange側でAPI呼ぶなら、同じことするだけなので無駄な気もする
-      //       そもそもonChange時にAPI通信した結果の反映がこのやり方で
-      //       正しい場合、refineの定義どうするかを考える
-      //
-      trigger('postalCode');
-      const zipcodeData = await getZipcodeOrList(
-        { zipcode: value },
-        { baseURL: 'https://zipcloud.ibsnet.co.jp' },
-      );
-      const {
-        status,
-        data: { results: data },
-      } = zipcodeData;
-      if (status === 200 && data !== null) {
-        setValue('prefectureId', data[0].prefcode);
-        setValue('town', data[0].address3);
-        //
-        // TODO: 外部APIの都合上、自動反映を後回し（watchしているのでリストは作成される）
-        //
-        // setValue('cityId', cityId);
-      }
-    }
-  };
-
-  //
-  // TODO: カスタムバリデーションを任意のタイミングで実行する方法を模索したい
-  //
-  const handleEmailChange = (value: string) => {
-    setValue('email', value);
-    trigger('email');
-  };
-
+  const {
+    control,
+    selectedYear,
+    selectedMonth,
+    selectedPrefecture,
+    handleYearChange,
+    handleMonthChange,
+    handleDayChange,
+    handlePostalCodeChange,
+    handleEmailChange,
+  } = useBasicInformaionForm();
   return (
     <>
       <div className='flex flex-row gap-x-4'>
@@ -205,7 +104,7 @@ export const BasicInfomationForm = () => {
                 <SelectTrigger ref={ref} className='w-[180px]'>
                   <SelectValue placeholder='西暦' />
                 </SelectTrigger>
-                <Year startYear={startYear} endYear={endYear} />
+                <Year />
                 <FormMessage />
               </Select>
             )}
